21 Dec 22

+++ ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: lde-sous <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/06 18:38:47 by lde-sous          #+#    #+#             */
-/*   Updated: 2022/12/15 12:22:19 by lde-sous         ###   ########.fr       */
+/*   Updated: 2022/12/20 17:04:55 by lde-sous         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int		ft_printf(const char *format, ...);
 char	ft_chck(char to_check);
 int		ft_printchar(char c);
 int		ft_printstring(char *str);
-int		ft_printhexa(unsigned long to_hex, char s);
+int		ft_printhexa(unsigned int to_hex, char s);
 int		ft_printnum(long num);
 int		ft_printpointer(unsigned long point);
 int		ft_printundec(unsigned int decimal);
